LayoutTests/fast/shapes/shape-inside/shape-inside-polygon-padding-002.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
<!--
The shape-inside polygon is just a square rotated 90 degrees. The X and Y offsets of the inset "padded"
boundary are padding * sqrt(2). To make the bouunds of the padded boundary work out to integral values
(as specified for the SVG polygon), the padding distance is 10 * sqrt(2) = 14.142, which makes the X and Y
offsets 20.
-->
<style>
    #shape-inside {
        position: relative;
        width: 500px;
        height: 500px;
        -webkit-shape-inside: polygon(300px 80px, 520px 300px, 300px 520px, 80px 300px);
        -webkit-shape-padding: 14.142px;
        font: 100px/1 Ahem, sans-serif;
        color: green;
    }

    #shape-outline {
        position: absolute;
        top: 0px;
        left: 0px;
        width: 500px;
        height: 500px;
    }
</style>
</head>
<body>
    <p>The blue diamond outline should be filled by cross made of 5 green squares.</p>
    <div id="shape-inside">
        <svg id="shape-outline" xmlns="http://www.w3.org/2000/svg">
            <polygon points="300,100 500,300 300,500 100,300" stroke="blue" fill="none"/>
        </svg>
    <span id="a">X</span> <span id="b">XXX</span> <span id="c">X</span>
    </div>
    <div id="console"></div>
</body>
<script>
  function shapeInsideRect(elementId) 
  {
      var s = document.getElementById("shape-inside").getBoundingClientRect();
      var r = document.getElementById(elementId).getBoundingClientRect();
      return {left: r.left - s.left, top: r.top - s.top, width: r.width, height: r.height};
  }

  var quiet = true; // PASS output depends on if subpixel layout has been enabled

  shouldBe("shapeInsideRect('a').top", "150");
  shouldBeCloseTo("shapeInsideRect('a').left", 250, 1, quiet);
  shouldBe("shapeInsideRect('a').width", "100");

  shouldBe("shapeInsideRect('b').top", "250");
  shouldBeCloseTo("shapeInsideRect('b').left", 150, 1, quiet)
  shouldBe("shapeInsideRect('b').width", "300");

  shouldBe("shapeInsideRect('c').top", "350"); 
  shouldBeCloseTo("shapeInsideRect('c').left", 250, 1, quiet);
  shouldBe("shapeInsideRect('c').width", "100"); 
</script>
</html>
